chg: dev: #223: address code review comments
